requirements-production.txt
# GENESIS Orchestrator - Production Requirements
# Pinned versions for reproducible builds
# Generated: 2025-01-14

# ==============================================================================
# CORE DEPENDENCIES
# ==============================================================================

# Temporal Workflow Engine
temporalio==1.4.0

# Data Validation & Serialization
pydantic==2.5.3
dataclasses-json==0.6.3

# Async HTTP
aiohttp==3.9.1
httpx==0.25.2

# ==============================================================================
# MONITORING & OBSERVABILITY (REQUIRED FOR PRODUCTION)
# ==============================================================================

# OpenTelemetry Core
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation==0.43b0

# OpenTelemetry Exporters
opentelemetry-exporter-prometheus==0.43b0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0

# Metrics
prometheus-client==0.19.0

# ==============================================================================
# DATABASE & CACHING
# ==============================================================================

# Redis
redis==5.0.1

# PostgreSQL
asyncpg==0.29.0

# ORM & Migrations
sqlalchemy==2.0.25
alembic==1.13.1

# MySQL (for Laravel backend compatibility)
PyMySQL==1.1.0

# ==============================================================================
# SECURITY (CRITICAL FOR PRODUCTION)
# ==============================================================================

# Cryptography
cryptography==41.0.7

# JWT Authentication
PyJWT==2.8.0

# Password Hashing
passlib[bcrypt]==1.7.4

# ==============================================================================
# TESTING (REQUIRED FOR CI/CD)
# ==============================================================================

# BDD Framework
behave==1.2.6

# Test Framework
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# ==============================================================================
# UTILITIES
# ==============================================================================

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1
environs==11.0.0

# JSON Processing
ujson==5.9.0
jsonlines==4.0.0
jsonschema==4.21.1

# Retry Logic
tenacity==8.2.3

# System Monitoring
psutil==5.9.8

# ==============================================================================
# LOGGING
# ==============================================================================

# Structured Logging
structlog==24.1.0

# ==============================================================================
# REQUIRED PYTHON VERSION
# ==============================================================================
# Python >= 3.9, < 3.12